all integrated

+++ src/main/java/tn/esprit/pidev/entities/evaluation/Score.java
@@ -4,8 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import tn.esprit.pidev.entities.user.User;
+import tn.esprit.pidev.entities.User;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -18,7 +17,7 @@ public class Score implements Serializable {
     @Id
     @GeneratedValue
     private Long numScore;
-    private int score;
+    private double score;
     private LocalDateTime dateTime;
 
     @ManyToOne
